bugFix: change dataType of HandleQRcmd()::separator_index from "size_t" to "unsigned int"
git-svn-id: https://unisim.org/svn/devel@12868 ea533459-3c0a-0410-a024-a90cf4b01ff5

+++ unisim_lib/unisim/service/debug/gdb_server/gdb_server.tcc
@@ -1376,7 +1376,8 @@ bool GDBServer<ADDRESS>::RemoveBreakpointWatchpoint(uint32_t type, ADDRESS addr,
 template <class ADDRESS>
 void GDBServer<ADDRESS>::HandleQRcmd(string command) {
 
-	size_t separator_index = command.find_first_of(':');
+//	size_t separator_index = command.find_first_of(':');
+	unsigned int separator_index = command.find_first_of(':');
 	string cmdPrefix;
 	if (separator_index == string::npos) {
 		cmdPrefix = command;
